Hardcode timezone to Americas/Los_Angeles in tests
Two of the tests were failing due to a weird timezone handling in 1893.

It was set to Americas/Tijuana, but this caused issues with 1893-04-01
as it doesn't map cleanly to an hour. In fact, I'm not really sure what
is going on with Tijuana back in 1893, as it would be 12 minutes earlier
than Los Angeles.

I'm using moment and moment-timezone below to illustrate. Weird. Would
love to know the answer.

```
> moment("1893-04-01").tz('America/Tijuana').format()
'1893-04-01T00:11:56-07:48'
> moment("1893-04-01").tz('America/Los_Angeles').format()
'1893-04-01T00:00:00-08:00'
```